+++ src/App.js
@@ -4,6 +4,9 @@ import React from 'react'
 // Library Imports
 import Header from 'modules/Header/Header'
 
+// Component Imports
+import SidebarLayout from 'layouts/SidebarLayout/SidebarLayout'
+
 // CSS Imports
 import GlobalStyle from 'src/globalConfig/GlobalStyles'
 import StyledApp from './StyledApp'
@@ -13,18 +16,7 @@ const App = () => {
     <StyledApp id='container'>
       <GlobalStyle />
       <Header id='headerWrap' />
-      <section id='sidebarWrap'>
-        <p>Sidebar:</p>
-      </section>
-      <section id='contentWrap'>
-        <h1>Welcome to React App thats build using Webpack and Babel separately</h1>
-        <h2>Heading 2</h2>
-        <h3>Heading 3</h3>
-        <h4>Heading 4</h4>
-        <h5>Heading 5</h5>
-        <h6>Heading 6</h6>
-        <p>Paragraph/body</p>
-      </section>
+      <SidebarLayout />
       <footer><p>&copy; 2023 - Karl Chvojka</p></footer>
     </StyledApp>
   )